14.08.2021 DEVAM2

+++ Gun5.Odev5_1.GameOrder/Entities/Campaign.cs
@@ -4,6 +4,6 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public double DiscountRate { get; set; }
+        public int DiscountRate { get; set; }
     }
 }
